Remove TaggedInstructions
