+++ src/riscv/lib/tests/expected/dummy/state_hash_final
@@ -1,1 +1,1 @@
-36714c44646704eee61cc79fff63322f1d796c08cb7bc55ed87e402d0bf2664c
+f0d5cd0d4630d0828227206b3f0957cee004e1f5517a0d83b07ddfbe38c3406c
