+++ src/riscv/lib/tests/expected/dummy/state_hash_initial
@@ -1,1 +1,1 @@
-4cbfe5c4c233ed94d1a8cf106e7f7cbc74ee2019e22aa8c4f57e6de6dbf3a313
+7ec2a511008bd686b762f82aac005cc8a19a27f7b29f09384a5035643e548883
