+++ src/riscv/lib/tests/expected/jstz/state_hash_final
@@ -1,1 +1,1 @@
-89939aa5f05a5b7c217efee5646793a6ff86f3e9f455103f4affab86b8f3d95f
+9c2ac8eb0037d84e631ab5c63f369da8f87431d5dbc62830cb22240ca3a4cd37
