+++ src/riscv/lib/tests/expected/jstz/state_hash_initial
@@ -1,1 +1,1 @@
-545deddecf5be00802308954dbbf76e52d5b524ad20dd79a40710f8fe9cc5886
+53e2b78071eb4319a0475cd89f84abfa90be603f5bf2638d4af11f02b2f4261c
